requirements.txt
Flask>=2.2.0
Werkzeug>=2.2.0
PyMuPDF>=1.19.0
Pillow>=9.0.0
pytesseract==0.3.8
fpdf>=1.7.2
gunicorn>=20.1.0
requests>=2.28.0
pytest>=7.0.0
pytest-cov>=3.0.0